// src/types/forms.ts
export enum NAME_FIELDS {
    fullName = "name",
    firstName = "firstName",
    lastName = "lastName",
    position = "position",
    company = "company",
    inn = "inn",
    kpp = "kpp",
    email = "email",
    phone = "phone",
    password = "password",
    oldPassword = "oldPassword",
    newPassword = "newPassword",
    confirmPassword = "confirmPassword",
    confirmNewPassword = "confirmNewPassword",
    code = "code",
    addressName = "name",
    city = "city",
    house = "house",
    street = "street",
    apartment = "apartment",
    isDefaultAddress = "isDefault",
    review = "review",
    rate = "rate",
    message = "message",
    agreement = "agreement",
}

export enum PLACEHOLDER_FIELDS {
    firstName = "Имя",
    firstNameReq = "Имя*",
    lastName = "Фамилия",
    lastNameReq = "Фамилия*",
    email = "Почта*",
    phone = "Телефон*",
    password = "Пароль*",
    oldPassword = "Старый пароль*",
    newPassword = "Новый пароль*",
    confirmPassword = "Подтверждение пароля*",
    confirmNewPassword = "Подтверждение нового пароля*",
    addressName = "Название адреса",
    city = "Город*",
    house = "Дом",
    street = "Улица*",
    apartment = "Квартира (офис)",
    company = "Компания*",
    inn = "ИНН*",
    kpp = "КПП*",
}